Refactored flex list factories Refactored overall projects structure
