+++ nice_droplets/dui.py
@@ -1,5 +1,3 @@
-import nice_droplets.components
-from .components.search_task import SearchTask as search_task
 from .elements.item import Item as item
 from .elements.item_section import ItemSection as item_section
 from .elements.item_label import ItemLabel as item_label
@@ -9,7 +7,6 @@ from .elements.typeahead import Typeahead as typeahead
 from .elements.flex_list import FlexList as flex_list
 
 __all__ = [
-    "components",
     "item",
     "item_section",
     "item_label",
@@ -17,5 +14,4 @@ __all__ = [
     "popover",
     "typeahead",
     "flex_list",
-    "search_task",
 ]
