+++ nice_droplets/elements/search_list.py
@@ -4,7 +4,8 @@ from nicegui.events import ValueChangeEventArguments, Handler
 from nice_droplets.components import SearchTask
 from nice_droplets.components.search_manager import SearchManager, SearchResultHandler
 from nice_droplets.events import SearchListContentUpdateEventArguments
-from nice_droplets.elements.flex_list import FlexList, FlexListFactory
+from nice_droplets.factories import FlexListFactory
+from nice_droplets.elements.flex_list import FlexList
