+++ nice_droplets/factories/__init__.py
@@ -1,0 +1,20 @@
+"""
+Factories for creating different types of lists.
+
+Available factories:
+- FlexLabelFactory: Simple list with label-based items
+- FlexItemListFactory: List with advanced item features (title, subtitle, avatar)
+- FlexTableFactory: Table-based list for structured data
+"""
+
+from .flex_list_factory import FlexListFactory
+from .flex_default_factory import FlexDefaultFactory
+from .flex_item_list_factory import FlexItemListFactory
+from .flex_table_factory import FlexTableFactory
+
+__all__ = [
+    'FlexListFactory',
+    'FlexDefaultFactory',
+    'FlexItemListFactory',
+    'FlexTableFactory',
+]
